Move default_mappings to test/support
This is not really a fixture and if we move this file the `/test`
directory will be much cleaner.
