+++ test/support/default_mappings.rb
